requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.5.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0
google-cloud-pubsub>=2.18.0
mysql-connector-python>=8.2.0
ariadne>=0.20.0
graphql-core>=3.2.0
